backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyTelegramBotAPI==4.18.0
python-dotenv==1.0.1
matplotlib==3.8.4
numpy==1.26.4
aiofiles==23.2.1
httpx==0.27.0
